fix(#790): break the ipeps_optimize <-> root_implicit cycle by hoisting shared leaves
test_algorithm_import_cycles_are_allowlisted_only has been red on main
since #773. Both edges were *function-local* imports, which is the
detail that made this look already-handled and stay broken:

  ipeps_optimize -> root_implicit  (dispatch, in a function body)
  root_implicit  -> ipeps_optimize (9 helpers, in function bodies)

Deferring an import dodges the *runtime* circular-import error, not the
cycle. tests/test_architecture_imports.py builds its graph with
ast.walk, so an import nested inside a function is an edge exactly like
a module-level one -- correctly, since the dependency is real either
way.

So the fix is not to move the imports around but to remove the reverse
edge entirely. The nine shared leaves move to _ipeps_optimize_shared,
which imports neither module:

  _build_optimizer          _normalize_params
  _converged_outer          _should_accept_best
  _grad_l2_norm             _use_line_search
  _log_ad_converged         _warn_implicit_ad_variational_caveat
  _wrap_as_dense_tensor

ipeps_optimize re-exports all nine, so the ~10 tests and benchmarks
doing `from tenax.algorithms.ipeps_optimize import _wrap_as_dense_tensor`
keep working -- verified, including object identity. ipeps_optimize
still dispatches *into* root_implicit; only the reverse edge is gone.

FlowDirection/TensorIndex/U1Symmetry drop out of ipeps_optimize with
_wrap_as_dense_tensor, after confirming nothing imports them through it.

Verification:
  test_architecture_imports.py   2 passed (was 1 failed)
  pytest -m core                 1486 passed, 8 skipped in 1:49:25
  ruff check / format            clean

The guard being red is the point: while it fails it cannot catch the
next cycle, and this epic keeps editing exactly this module pair.

Refs #790, #715. Related: #805 -- both files sit in _UNBUCKETED_LEGACY,
which is why this ran red in no required CI job.
